fetching session products
