+++ web/src/pages/cart/index.tsx
@@ -4,8 +4,8 @@ import { useEffect, useState } from 'react'
 import { useSelector } from 'react-redux'
 import { selectCartItems, selectCartTotal } from '@redux/cartSlice'
 import { Stripe } from 'stripe'
-import getStripe from 'utils/getStripe'
-import { fetchPostJSON } from 'helpers/fetchPostJSON'
+import getStripe from '@utils/getStripe'
+import { fetchPostJSON } from '@utils/fetchPostJSON'
 import Divider from '@components/Divider'
 import Headline from '@components/Headline'
 import Button from '@components/Button'
